+++ src/Projects/Coree.NETStandard/CoreeHttpClient/CoreeHttpClient2.cs
@@ -3,8 +3,8 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Threading.Tasks;
 using System.Threading;
+using System.Threading.Tasks;
 
 using Json.Path;
