+++ src/Projects/Coree.NETStandard/Services/Process/DSProcessService.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.Text;
 
 using Coree.NETStandard.Abstractions;
+using Coree.NETStandard.Abstractions.DependencySingleton;
 using Coree.NETStandard.Services.File;
+
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
