+++ src/Projects/Coree.NETStandard/Services/RuntimeInsights/DSRuntimeInsightsService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 
 using Coree.NETStandard.Abstractions;
+using Coree.NETStandard.Abstractions.DependencySingleton;
 
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
